fix: rename some v0_methods to v1_methods
