+++ packages/rs-dpp/src/state_transition/state_transitions/document/batch_transition/v1/mod.rs
@@ -3,7 +3,7 @@ mod identity_signed;
 mod json_conversion;
 mod state_transition_like;
 mod types;
-mod v0_methods;
+mod v1_methods;
 #[cfg(feature = "state-transition-value-conversion")]
 mod value_conversion;
 mod version;
