+++ packages/rs-dpp/src/state_transition/state_transitions/document/batch_transition/v1/v1_methods.rs
